requirements.txt
fastapi
uvicorn
langchain
ollama
requests
python-dateutil
pydantic
python-dotenv
tiktoken
